UI : create data binding between recyclerview binding and its item

+++ app/src/main/java/com/jiwon/prestolabs/model/Data.kt
@@ -9,7 +9,7 @@ data class Instrument(
     @SerializedName("isInverse") var isInverse : Boolean,
     @SerializedName("lastPrice") var lastPrice : Int = 0,
     @SerializedName("lastChangePcnt") var lastChangePercentage : Double = 0.0,
-    @SerializedName("volume24h") var volume24 : Int = 0
+    @SerializedName("volume24h") var volume24 : Long = 0
 ){
     override fun toString(): String {
         return "symbol : $symbol, state : $state, isInverse : ${isInverse}"
@@ -47,7 +47,7 @@ data class InstrumentUpdate(
     @SerializedName("symbol") val symbol : String,
     @SerializedName("lastPrice") var lastPrice : Int = 0,
     @SerializedName("lastChangePcnt") var lastChangePercentage : Double = 0.0,
-    @SerializedName("volume24h") var volume24 : Int = 0
+    @SerializedName("volume24h") var volume24 : Long = 0
 ){
     override fun toString(): String {
         return "symbol : $symbol, lastprice : $lastPrice, lastChangePercentage : ${lastChangePercentage} volume24 : $volume24"
